# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  http.OrganResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      slug:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  http.OrganUpsertRequest:
    properties:
      description:
        example: Everything related to money, cashflow, budgeting
        type: string
      slug:
        example: finance
        type: string
      title:
        example: Finance
        type: string
    required:
    - slug
    - title
    type: object
  http.PlanResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      diseaseId:
        type: string
      id:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  http.PlanUpsertRequest:
    properties:
      description:
        example: 7-day actions to stabilize cashflow
        type: string
      title:
        example: Cashflow recovery plan
        type: string
    required:
    - title
    type: object
  http.PlanWithStepsResponse:
    properties:
      plan:
        $ref: '#/definitions/http.PlanResponse'
      steps:
        items:
          $ref: '#/definitions/http.StepResponse'
        type: array
    type: object
...
  http.StepResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      orderNo:
        type: integer
      planId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  http.updateDashboardUserReq:
    properties:
      name:
        type: string
      surname:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Business Surgery service.
  title: Business Surgery Backend API
  version: "1.0"
paths:
...
  /api/v1/diseases/{id}/plan:
    get:
      parameters:
      - description: Disease ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PlanWithStepsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get plan by disease id (with steps)
      tags:
      - plans
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Disease ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Plan payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.PlanUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PlanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create/update plan for disease
      tags:
      - plans
...
  /api/v1/moderator/dashboard/users/{id}:
    delete:
      description: Deletes participant from moderator dashboard
      parameters:
      - description: User ID (uuid)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete participant
      tags:
      - moderator-dashboard
    put:
      consumes:
      - application/json
      description: Updates participant name and surname from moderator dashboard
      parameters:
      - description: User ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateDashboardUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update participant profile
      tags:
      - moderator-dashboard
...
  /api/v1/organs:
    get:
      description: Returns list of business organs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.OrganResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List organs
      tags:
      - organs
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organ payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.OrganUpsertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.OrganResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create organ
      tags:
      - organs
  /api/v1/organs/{id}:
    delete:
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organ ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete organ
      tags:
      - organs
    get:
      parameters:
      - description: Organ ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.OrganResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get organ by id
      tags:
      - organs
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organ ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: Organ payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.OrganUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.OrganResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update organ
      tags:
      - organs
  /api/v1/plans/{planId}:
    get:
      parameters:
      - description: Plan ID (uuid)
        in: path
        name: planId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PlanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get plan by id
      tags:
      - plans
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"